+++ src/docs/animate/index.html
@@ -13,7 +13,7 @@
 	animation-keyframe-from-transform="scale(1)"
 	animation-keyframe-50-transform="scale(0)"
 	animation-keyframe-to-transform="scale(1)"
-	animation-option-duration="700"
+	animation-option-duration="1000"
 	animation-option-delay="100"
 	animation-option-easing="ease-in"
 >
