+++ src/docs/copy/index.html
@@ -1,6 +1,6 @@
 <script type="module" src="/package/copy/define.ts"></script>
 
-<drab-copy text="Text to copy" complete="Copied">
+<drab-copy value="Text to copy" complete="Copied">
 	<button data-trigger data-content type="button" class="button button-primary">
 		Copy Text
 	</button>
