+++ src/docs/share/index.html
@@ -1,6 +1,6 @@
 <script type="module" src="/package/share/define.ts"></script>
 
-<drab-share url="https://robino.dev" complete="Copied">
+<drab-share value="https://robino.dev" complete="Copied">
 	<button data-trigger data-content type="button" class="button button-primary">
 		Share URL
 	</button>
