+++ src/lib/docs/classes/base.md
@@ -15,11 +15,7 @@ default selector from `[data-trigger]` and `[data-content]`.
 
   ↳↳ [`Animate`](/docs/classes/Animate.md)
 
-  ↳↳ [`Copy`](/docs/classes/Copy.md)
-
   ↳↳ [`Fullscreen`](/docs/classes/Fullscreen.md)
-
-  ↳↳ [`Share`](/docs/classes/Share.md)
 
   ↳↳ [`YouTube`](/docs/classes/YouTube.md)
 
@@ -41,7 +37,7 @@ HTMLElement.constructor
 
 #### Defined in
 
-[src/package/base/index.ts:16](https://github.com/rossrobino/components/blob/3b25f8c/src/package/base/index.ts#L16)
+[src/package/base/index.ts:16](https://github.com/rossrobino/components/blob/8918ae1/src/package/base/index.ts#L16)
 
 ---
 
@@ -56,7 +52,7 @@ when the element is removed.
 
 #### Defined in
 
-[src/package/base/index.ts:14](https://github.com/rossrobino/components/blob/3b25f8c/src/package/base/index.ts#L14)
+[src/package/base/index.ts:14](https://github.com/rossrobino/components/blob/8918ae1/src/package/base/index.ts#L14)
 
 ---
 
@@ -80,7 +76,7 @@ keyof `HTMLElementEventMap`
 
 #### Defined in
 
-[src/package/base/index.ts:25](https://github.com/rossrobino/components/blob/3b25f8c/src/package/base/index.ts#L25)
+[src/package/base/index.ts:25](https://github.com/rossrobino/components/blob/8918ae1/src/package/base/index.ts#L25)
 
 • `set` **triggerEvent**(`value`): `void`
 
@@ -96,7 +92,7 @@ keyof `HTMLElementEventMap`
 
 #### Defined in
 
-[src/package/base/index.ts:30](https://github.com/rossrobino/components/blob/3b25f8c/src/package/base/index.ts#L30)
+[src/package/base/index.ts:30](https://github.com/rossrobino/components/blob/8918ae1/src/package/base/index.ts#L30)
 
 ---
 
@@ -132,7 +128,7 @@ this.querySelector("[data-content]");
 
 #### Defined in
 
-[src/package/base/index.ts:52](https://github.com/rossrobino/components/blob/3b25f8c/src/package/base/index.ts#L52)
+[src/package/base/index.ts:52](https://github.com/rossrobino/components/blob/8918ae1/src/package/base/index.ts#L52)
 
 ### disconnectedCallback
 
@@ -144,7 +140,7 @@ this.querySelector("[data-content]");
 
 #### Defined in
 
-[src/package/base/index.ts:77](https://github.com/rossrobino/components/blob/3b25f8c/src/package/base/index.ts#L77)
+[src/package/base/index.ts:77](https://github.com/rossrobino/components/blob/8918ae1/src/package/base/index.ts#L77)
 
 ### safeAddEventListener
 
@@ -173,7 +169,7 @@ element is removed from the DOM, these event listeners are cleaned up.
 
 #### Defined in
 
-[src/package/base/index.ts:68](https://github.com/rossrobino/components/blob/3b25f8c/src/package/base/index.ts#L68)
+[src/package/base/index.ts:68](https://github.com/rossrobino/components/blob/8918ae1/src/package/base/index.ts#L68)
 
 ### trigger
 
@@ -193,4 +189,4 @@ this.querySelectorAll("[data-trigger]");
 
 #### Defined in
 
-[src/package/base/index.ts:38](https://github.com/rossrobino/components/blob/3b25f8c/src/package/base/index.ts#L38)
+[src/package/base/index.ts:38](https://github.com/rossrobino/components/blob/8918ae1/src/package/base/index.ts#L38)
