+++ src/lib/docs/classes/details.md
@@ -43,7 +43,7 @@ Animations options can also be set:
 
 #### Defined in
 
-[src/package/details/index.ts:4](https://github.com/rossrobino/components/blob/3b25f8c/src/package/details/index.ts#L4)
+[src/package/details/index.ts:4](https://github.com/rossrobino/components/blob/8918ae1/src/package/details/index.ts#L4)
 
 ---
 
@@ -62,7 +62,7 @@ when the element is removed.
 
 #### Defined in
 
-[src/package/base/index.ts:14](https://github.com/rossrobino/components/blob/3b25f8c/src/package/base/index.ts#L14)
+[src/package/base/index.ts:14](https://github.com/rossrobino/components/blob/8918ae1/src/package/base/index.ts#L14)
 
 ---
 
@@ -84,7 +84,7 @@ Animate.animationOptions
 
 #### Defined in
 
-[src/package/animate/index.ts:32](https://github.com/rossrobino/components/blob/3b25f8c/src/package/animate/index.ts#L32)
+[src/package/animate/index.ts:32](https://github.com/rossrobino/components/blob/8918ae1/src/package/animate/index.ts#L32)
 
 ### details
 
@@ -96,7 +96,7 @@ Animate.animationOptions
 
 #### Defined in
 
-[src/package/details/index.ts:8](https://github.com/rossrobino/components/blob/3b25f8c/src/package/details/index.ts#L8)
+[src/package/details/index.ts:8](https://github.com/rossrobino/components/blob/8918ae1/src/package/details/index.ts#L8)
 
 ### keyframes
 
@@ -112,7 +112,7 @@ Animate.keyframes
 
 #### Defined in
 
-[src/package/animate/index.ts:108](https://github.com/rossrobino/components/blob/3b25f8c/src/package/animate/index.ts#L108)
+[src/package/animate/index.ts:108](https://github.com/rossrobino/components/blob/8918ae1/src/package/animate/index.ts#L108)
 
 ### triggerEvent
 
@@ -136,7 +136,7 @@ Animate.triggerEvent
 
 #### Defined in
 
-[src/package/base/index.ts:25](https://github.com/rossrobino/components/blob/3b25f8c/src/package/base/index.ts#L25)
+[src/package/base/index.ts:25](https://github.com/rossrobino/components/blob/8918ae1/src/package/base/index.ts#L25)
 
 • `set` **triggerEvent**(`value`): `void`
 
@@ -156,7 +156,7 @@ Animate.triggerEvent
 
 #### Defined in
 
-[src/package/base/index.ts:30](https://github.com/rossrobino/components/blob/3b25f8c/src/package/base/index.ts#L30)
+[src/package/base/index.ts:30](https://github.com/rossrobino/components/blob/8918ae1/src/package/base/index.ts#L30)
 
 ---
 
@@ -193,7 +193,7 @@ Animates a particular element using the web animations API.
 
 #### Defined in
 
-[src/package/animate/index.ts:63](https://github.com/rossrobino/components/blob/3b25f8c/src/package/animate/index.ts#L63)
+[src/package/animate/index.ts:63](https://github.com/rossrobino/components/blob/8918ae1/src/package/animate/index.ts#L63)
 
 ### close
 
@@ -207,7 +207,7 @@ Closes details with animation.
 
 #### Defined in
 
-[src/package/details/index.ts:21](https://github.com/rossrobino/components/blob/3b25f8c/src/package/details/index.ts#L21)
+[src/package/details/index.ts:21](https://github.com/rossrobino/components/blob/8918ae1/src/package/details/index.ts#L21)
 
 ### connectedCallback
 
@@ -219,7 +219,7 @@ Closes details with animation.
 
 #### Defined in
 
-[src/package/details/index.ts:36](https://github.com/rossrobino/components/blob/3b25f8c/src/package/details/index.ts#L36)
+[src/package/details/index.ts:36](https://github.com/rossrobino/components/blob/8918ae1/src/package/details/index.ts#L36)
 
 ### content
 
@@ -255,7 +255,7 @@ this.querySelector("[data-content]");
 
 #### Defined in
 
-[src/package/base/index.ts:52](https://github.com/rossrobino/components/blob/3b25f8c/src/package/base/index.ts#L52)
+[src/package/base/index.ts:52](https://github.com/rossrobino/components/blob/8918ae1/src/package/base/index.ts#L52)
 
 ### disconnectedCallback
 
@@ -271,7 +271,7 @@ this.querySelector("[data-content]");
 
 #### Defined in
 
-[src/package/base/index.ts:77](https://github.com/rossrobino/components/blob/3b25f8c/src/package/base/index.ts#L77)
+[src/package/base/index.ts:77](https://github.com/rossrobino/components/blob/8918ae1/src/package/base/index.ts#L77)
 
 ### open
 
@@ -283,7 +283,7 @@ this.querySelector("[data-content]");
 
 #### Defined in
 
-[src/package/details/index.ts:15](https://github.com/rossrobino/components/blob/3b25f8c/src/package/details/index.ts#L15)
+[src/package/details/index.ts:15](https://github.com/rossrobino/components/blob/8918ae1/src/package/details/index.ts#L15)
 
 ### safeAddEventListener
 
@@ -316,7 +316,7 @@ element is removed from the DOM, these event listeners are cleaned up.
 
 #### Defined in
 
-[src/package/base/index.ts:68](https://github.com/rossrobino/components/blob/3b25f8c/src/package/base/index.ts#L68)
+[src/package/base/index.ts:68](https://github.com/rossrobino/components/blob/8918ae1/src/package/base/index.ts#L68)
 
 ### toggle
 
@@ -328,7 +328,7 @@ element is removed from the DOM, these event listeners are cleaned up.
 
 #### Defined in
 
-[src/package/details/index.ts:28](https://github.com/rossrobino/components/blob/3b25f8c/src/package/details/index.ts#L28)
+[src/package/details/index.ts:28](https://github.com/rossrobino/components/blob/8918ae1/src/package/details/index.ts#L28)
 
 ### trigger
 
@@ -352,4 +352,4 @@ this.querySelectorAll("[data-trigger]");
 
 #### Defined in
 
-[src/package/base/index.ts:38](https://github.com/rossrobino/components/blob/3b25f8c/src/package/base/index.ts#L38)
+[src/package/base/index.ts:38](https://github.com/rossrobino/components/blob/8918ae1/src/package/base/index.ts#L38)
